pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llms2"
version = "0.1.0"
description = "A Python package for managing multiple LLM profiles and API interactions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]
